+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - css bug incompatible with url perfix
-
+- fix event time to UTC conversion.
 
 ## [1.0.2] - 2019-10-15
 ### Fixed